move content type encoding and decoding to contenttype package
