+++ connector/internal/contenttype/data_uri.go
@@ -1,4 +1,4 @@
-package internal
+package contenttype
 
 import (
 	"encoding/base64"
