+++ connector/internal/contenttype/xml_encode_test.go
@@ -1,4 +1,4 @@
-package internal
+package contenttype
 
 import (
 	"bytes"
