+++ connector/internal/utils.go
@@ -4,11 +4,8 @@ import (
 	"fmt"
 	"net/http"
 	"net/url"
-	"reflect"
-	"strconv"
 	"strings"
 
-	rest "github.com/hasura/ndc-http/ndc-http-schema/schema"
 	"github.com/hasura/ndc-http/ndc-http-schema/utils"
 	"github.com/hasura/ndc-sdk-go/schema"
 	"go.opentelemetry.io/otel/attribute"
@@ -61,82 +58,5 @@ func cloneURL(input *url.URL) *url.URL {
 		RawQuery:    input.RawQuery,
 		Fragment:    input.Fragment,
 		RawFragment: input.RawFragment,
-	}
-}
-
-func stringifySimpleScalar(val reflect.Value, kind reflect.Kind) (string, error) {
-	switch kind {
-	case reflect.Int, reflect.Int8, reflect.Int16, reflect.Int32, reflect.Int64:
-		return strconv.FormatInt(val.Int(), 10), nil
-	case reflect.Uint, reflect.Uint8, reflect.Uint16, reflect.Uint32, reflect.Uint64, reflect.Uintptr:
-		return strconv.FormatUint(val.Uint(), 10), nil
-	case reflect.Float32, reflect.Float64:
-		return strconv.FormatFloat(val.Float(), 'g', -1, val.Type().Bits()), nil
-	case reflect.String:
-		return val.String(), nil
-	case reflect.Bool:
-		return strconv.FormatBool(val.Bool()), nil
-	case reflect.Interface:
-		return fmt.Sprint(val.Interface()), nil
-	default:
-		return "", fmt.Errorf("invalid value: %v", val.Interface())
-	}
-}
-
-func findXMLLeafObjectField(objectType rest.ObjectType) (*rest.ObjectField, string, bool) {
-	var f *rest.ObjectField
-	var fieldName string
-	for key, field := range objectType.Fields {
-		if field.HTTP == nil || field.HTTP.XML == nil {
-			return nil, "", false
-		}
-		if field.HTTP.XML.Text {
-			f = &field
-			fieldName = key
-		} else if !field.HTTP.XML.Attribute {
-			return nil, "", false
-		}
-	}
-
-	return f, fieldName, true
-}
-
-func getTypeSchemaXMLName(typeSchema *rest.TypeSchema, defaultName string) string {
-	if typeSchema != nil {
-		return getXMLName(typeSchema.XML, defaultName)
-	}
-
-	return defaultName
-}
-
-func getXMLName(xmlSchema *rest.XMLSchema, defaultName string) string {
-	if xmlSchema != nil {
-		if xmlSchema.Name != "" {
-			return xmlSchema.GetFullName()
-		}
-
-		if xmlSchema.Prefix != "" {
-			return xmlSchema.Prefix + ":" + defaultName
-		}
-	}
-
-	return defaultName
-}
-
-func getArrayOrNamedType(schemaType schema.Type) (*schema.ArrayType, *schema.NamedType, error) {
-	rawType, err := schemaType.InterfaceT()
-	if err != nil {
-		return nil, nil, err
-	}
-
-	switch t := rawType.(type) {
-	case *schema.NullableType:
-		return getArrayOrNamedType(t.UnderlyingType)
-	case *schema.ArrayType:
-		return t, nil, nil
-	case *schema.NamedType:
-		return nil, t, nil
-	default:
-		return nil, nil, nil
 	}
 }
